requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
openai
tqdm
pandas

# RodionfromHSE
accelerate==0.24.1
aiohttp==3.8.6
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.3.1
colorama==0.4.6
comm==0.1.4
datasets==2.14.6
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
executing==2.0.0
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.10.0
huggingface-hub==0.17.3
hydra-core==1.3.2
idna==3.4
ipykernel==6.26.0
ipython==8.16.1
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.5.0
jupyter_core==5.4.0
jupyterlab-widgets==3.0.9
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.1
omegaconf==2.3.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
portalocker==2.8.2
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
sacrebleu==2.3.2
safetensors==0.4.0
sentencepiece==0.1.99
six==1.16.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tokenizers==0.14.1
torch==2.1.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
transformers==4.34.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.8
widgetsnbextension==4.0.9
xxhash==3.4.1
yarl==1.9.2
zstandard==0.21.0
